+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.MVVM1"
         tools:targetApi="31" >
         <activity
-            android:name=".MainActivity"
+            android:name=".View.MainActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
