--- app/src/main/java/com/mariovaladez/mvvm1/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.mariovaladez.mvvm1
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.mariovaladez.mvvm1.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding:ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-    }
-}
