+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".MainActivity">
+    tools:context=".View.MainActivity">
     <TextView
         android:id="@+id/tvQuote"
       android:layout_width="wrap_content"
